+++ src/main/java/hr/bp/adventofcode/day06/RaceRecord.java
@@ -1,8 +1,8 @@
 package hr.bp.adventofcode.day06;
 
-public record RaceRecord(int maximumTime, int recordDistance) {
+public record RaceRecord(long maximumTime, long recordDistance) {
 
     public RaceRecord(String maximumTime, String distanceRecord) {
-        this(Integer.parseInt(maximumTime), Integer.parseInt(distanceRecord));
+        this(Long.parseLong(maximumTime), Long.parseLong(distanceRecord));
     }
 }
